refactored some components of edit business, and also added get current location, will be testing it now
