+++ android/app/src/main/AndroidManifest.xml
@@ -7,6 +7,8 @@
             <data android:scheme="tel" />
         </intent>
     </queries>
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
    <application
         android:label="gomart"
         android:name="${applicationName}"
