+++ ios/Runner/Info.plist
@@ -50,6 +50,8 @@
 	<string>Gomart uses a feature that requires user to capture photo with his camera and upload it</string>
 	<key>NSPhotoLibraryUsageDescription</key>
 	<string>Gomart uses has features that require user to upload his photo or a product photo</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>Gomart uses location when app is in use and when we need to use your location to suggest nearby products & Services</string>
 	<key>UIBackgroundModes</key>
 	<array>
 		<string>remote-notification</string>
