+++ lib/screens/Home/Screens/Account/ManageProduct/manage_product_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
-import '../../../../styles/custom_home_icons.dart';
-import '../../../../styles/styles.dart';
-import 'business_profile_screen.dart';
-import 'EditBusiness/edit_business_screen.dart';
+import '../../../../../styles/custom_home_icons.dart';
+import '../../../../../styles/styles.dart';
+import '../BusinessProfile/business_profile_screen.dart';
+import '../EditBusiness/_text_field.dart';
 
 class ManageProductScreen extends StatelessWidget {
   const ManageProductScreen({Key? key}) : super(key: key);
