+++ lib/shared_components/utilities/_horizontal_pad.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class HorizontalPad extends StatelessWidget {
+  final Widget child;
+  final double padding;
+
+  const HorizontalPad({super.key, required this.child, this.padding = 8.0});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.symmetric(horizontal: padding),
+      child: child,
+    );
+  }
+}
